change Long to BigInteger in ede service

+++ enrolment-deferment-exemption/src/main/java/com/microservice/erp/domain/dto/ParentConsentListDto.java
@@ -3,10 +3,12 @@ package com.microservice.erp.domain.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Setter
 @Getter
 public class ParentConsentListDto {
-    private Long userId;
+    private BigInteger userId;
     private String fullName;
     private String guardianName;
     private String guardianMobileNo;
