+++ enrolment-deferment-exemption/src/main/java/com/microservice/erp/domain/dto/parentConsent/ParentConsentDto.java
@@ -3,10 +3,12 @@ package com.microservice.erp.domain.dto.parentConsent;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Setter
 @Getter
 public class ParentConsentDto {
-    private Long userId;
+    private BigInteger userId;
     private String mobileNo;
     private String fullName;
     private String email;
