+++ enrolment-deferment-exemption/src/main/java/com/microservice/erp/domain/entities/EnrolmentCoursePreference.java
@@ -6,19 +6,20 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigInteger;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity(name = "ede_enrolment_course_preference")
 @AttributeOverride(name = "id", column = @Column(name = "enrolment_course_preference_id"))
-public class EnrolmentCoursePreference extends Auditable<Long, Long> {
+public class EnrolmentCoursePreference extends Auditable<BigInteger, Long> {
 
     //region private variables
     @Column(name = "course_id")
-    private Long courseId;
+    private BigInteger courseId;
 
     @Column(name = "user_id")
-    private Long userId;
+    private BigInteger userId;
 
     @NotNull(message = "Priority number cannot not be null")
     @Column(name = "preference_number")
@@ -31,19 +32,19 @@ public class EnrolmentCoursePreference extends Auditable<Long, Long> {
 
     //region setters and getters
 
-    public Long getCourseId() {
+    public BigInteger getCourseId() {
         return courseId;
     }
 
-    public void setCourseId(Long courseId) {
+    public void setCourseId(BigInteger courseId) {
         this.courseId = courseId;
     }
 
-    public Long getUserId() {
+    public BigInteger getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(BigInteger userId) {
         this.userId = userId;
     }
