+++ enrolment-deferment-exemption/src/main/java/com/microservice/erp/services/iServices/deferment/IUpdateDefermentService.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 import org.springframework.http.ResponseEntity;
 
 import javax.validation.Valid;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IUpdateDefermentService {
@@ -22,6 +23,6 @@ public interface IUpdateDefermentService {
 
         private String remarks;
 
-        private List<Long> defermentIds;
+        private List<BigInteger> defermentIds;
     }
 }
