+++ enrolment-deferment-exemption/src/main/java/com/microservice/erp/services/impl/deferment/UpdateDefermentService.java
@@ -19,6 +19,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
 import javax.validation.Valid;
+import java.math.BigInteger;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
@@ -86,7 +87,7 @@ public class UpdateDefermentService implements IUpdateDefermentService {
 
     }
 
-    private void sendEmailAndSms(String authHeader, Long userId, Character status) throws Exception {
+    private void sendEmailAndSms(String authHeader, BigInteger userId, Character status) throws Exception {
         RestTemplate restTemplate = new RestTemplate();
         HttpEntity<String> httpRequest = headerToken.tokenHeader(authHeader);
